+++ packages/perseus-core/src/parse-perseus-json/regression-tests/parse-perseus-json-regression.test.ts
@@ -1,20 +1,65 @@
 import * as fs from "fs";
 import {join} from "path";
 
-import {parseAndMigratePerseusItem} from "../index";
+import {anySuccess} from "../general-purpose-parsers/test-helpers";
+import {
+    parseAndMigratePerseusArticle,
+    parseAndMigratePerseusItem,
+} from "../index";
+import {assertSuccess, mapFailure} from "../result";
 
-const dataFiles = fs.readdirSync(join(__dirname, "data"));
+const itemDataDir = join(__dirname, "item-data");
+const itemDataFiles = fs.readdirSync(itemDataDir);
 
-describe("parseAndTypecheckPerseusItem", () => {
-    it.each(dataFiles)("accepts data/%s", (filename) => {
-        const json = fs.readFileSync(
-            join(__dirname, "data", filename),
-            "utf-8",
-        );
+const articleDataDir = join(__dirname, "article-data");
+const articleDataFiles = fs.readdirSync(articleDataDir);
+
+describe("parseAndMigratePerseusItem", () => {
+    describe.each(itemDataFiles)("given %s", (filename) => {
+        const json = fs.readFileSync(join(itemDataDir, filename), "utf-8");
         const result = parseAndMigratePerseusItem(json);
 
-        // This strange-looking assertion style results in the failure message
-        // being printed if parsing fails, so the test is easier to debug.
-        expect(result).toEqual(expect.objectContaining({type: "success"}));
+        it("parses successfully", () => {
+            // If the parse fails, get just the error message. This makes the test
+            // failure easier to read, since otherwise the entire `invalidObject`
+            // from the ParseFailureDetail would be printed.
+            const result = mapFailure(getMessage)(
+                parseAndMigratePerseusItem(json),
+            );
+
+            expect(result).toEqual(anySuccess);
+        });
+
+        it("returns the same result as before", () => {
+            assertSuccess(result);
+            expect(result.value).toMatchSnapshot();
+        });
     });
 });
+
+describe("parseAndMigratePerseusArticle", () => {
+    describe.each(articleDataFiles)("given %s", (filename) => {
+        const json = fs.readFileSync(join(articleDataDir, filename), "utf-8");
+        const result = parseAndMigratePerseusArticle(json);
+
+        it("parses successfully", () => {
+            // If the parse fails, get just the error message. This makes the test
+            // failure easier to read, since otherwise the entire `invalidObject`
+            // from the ParseFailureDetail would be printed.
+            const result = mapFailure(getMessage)(
+                parseAndMigratePerseusArticle(json),
+            );
+
+            expect(result).toEqual(anySuccess);
+        });
+
+        it("returns the same result as before", () => {
+            assertSuccess(result);
+            expect(result.value).toMatchSnapshot();
+        });
+    });
+});
+
+function getMessage(obj: {message: string}): string {
+    return obj.message;
+}