+++ packages/perseus-core/src/widgets/sorter/sorter-util.ts
@@ -1,3 +1,5 @@
+import {shuffle} from "@khanacademy/perseus-core";
+
 import type {PerseusSorterWidgetOptions} from "@khanacademy/perseus-core";
 
 /**
@@ -8,6 +10,7 @@ type SorterPublicWidgetOptions = {
     correct: PerseusSorterWidgetOptions["correct"];
     padding: PerseusSorterWidgetOptions["padding"];
     layout: PerseusSorterWidgetOptions["layout"];
+    isCorrectShuffled: boolean;
 };
 
 /**
@@ -17,13 +20,22 @@ type SorterPublicWidgetOptions = {
 function getSorterPublicWidgetOptions(
     options: PerseusSorterWidgetOptions,
 ): SorterPublicWidgetOptions {
+    const shuffledCorrect = shuffle(
+        options.correct,
+        Math.random,
+        /* ensurePermuted */ true,
+    );
+
     return {
+        ...options,
         // Note(Tamara): This does not provide correct answer information any longer.
         // To maintain compatibility with the original widget options, we are
         // keeping the key the same. Represents initial state of the cards here.
-        correct: options.correct.slice().sort(),
-        padding: options.padding,
-        layout: options.layout,
+        correct: shuffledCorrect,
+        // Note(Tamara): This new key is only added here with "true". There isn't
+        // a place where it is set to false. It indicates that the correct field
+        // has been shuffled and no longer contains correct answer info.
+        isCorrectShuffled: true,
     };
 }