+++ packages/perseus/src/widgets/dropdown/dropdown.tsx
@@ -1,4 +1,3 @@
-import {getDropdownPublicWidgetOptions} from "@khanacademy/perseus-core";
 import {Id, View} from "@khanacademy/wonder-blocks-core";
 import {SingleSelect, OptionItem} from "@khanacademy/wonder-blocks-dropdown";
 import {LabelLarge} from "@khanacademy/wonder-blocks-typography";
@@ -167,5 +166,4 @@ export default {
     accessible: true,
     widget: Dropdown,
     transform,
-    getPublicWidgetOptions: getDropdownPublicWidgetOptions,
 } satisfies WidgetExports<typeof Dropdown>;